Move widget related classes out of their impl package
